fix(UVE): create DotUVENgEvenHandlerService
