+++ core-web/libs/portlets/edit-ema/portlet/src/lib/dot-ema-shell/dot-ema-shell.component.html
@@ -1,5 +1,5 @@
 @if ($shellProps()?.canRead) {
-    <router-outle />
+    <router-outlet />
     <dot-edit-ema-navigation-bar
         data-testId="ema-nav-bar"
         [items]="$shellProps().items"
