+++ core-web/libs/portlets/edit-ema/portlet/src/lib/dot-ema-shell/dot-ema-shell.component.ts
@@ -1,8 +1,5 @@
-import { EMPTY } from 'rxjs';
-
 import { CommonModule, Location } from '@angular/common';
-import { HttpErrorResponse } from '@angular/common/http';
-import { ChangeDetectorRef, Component, effect, inject, OnInit, ViewChild } from '@angular/core';
+import { Component, effect, inject, OnInit, ViewChild } from '@angular/core';
 import { ActivatedRoute, Params, Router, RouterModule } from '@angular/router';
 
 import { ConfirmationService, MessageService } from 'primeng/api';
@@ -10,9 +7,8 @@ import { ConfirmDialogModule } from 'primeng/confirmdialog';
 import { DialogService } from 'primeng/dynamicdialog';
 import { ToastModule } from 'primeng/toast';
 
-import { catchError, filter, map, skip, take, tap } from 'rxjs/operators';
+import { skip } from 'rxjs/operators';
 
-import { CLIENT_ACTIONS } from '@dotcms/client';
 import {
     DotESContentService,
     DotExperimentsService,
@@ -23,10 +19,7 @@ import {
     DotSeoMetaTagsService,
     DotSeoMetaTagsUtilService,
     DotWorkflowsActionsService,
-    DotAnalyticsTrackerService,
-    DotMessageService,
-    DotContentletService,
-    DotHttpErrorManagerService
+    DotAnalyticsTrackerService
 } from '@dotcms/data-access';
 import { SiteService } from '@dotcms/dotcms-js';
 import { DotPageToolsSeoComponent } from '@dotcms/portlets/dot-ema/ui';
@@ -39,19 +32,16 @@ import { EditEmaNavigationBarComponent } from './components/edit-ema-navigation-
 import { DotEmaDialogComponent } from '../components/dot-ema-dialog/dot-ema-dialog.component';
 import { DotEditorDialogService } from '../components/dot-ema-dialog/services/dot-ema-dialog.service';
 import { DotActionUrlService } from '../services/dot-action-url/dot-action-url.service';
+import { DotUVENgEvenHandlerService } from '../services/dot-ng-event-handler/dot-ng-event-handler.service';
 import { DotPageApiService } from '../services/dot-page-api.service';
-import { EDITOR_STATE, NG_CUSTOM_EVENTS, UVE_STATUS } from '../shared/enums';
 import { DialogAction, DotPageAssetParams } from '../shared/models';
 import { UVEStore } from '../store/dot-uve.store';
 import { DotUveViewParams } from '../store/models';
 import {
     checkClientHostAccess,
     getAllowedPageParams,
-    getTargetUrl,
-    insertContentletInContainer,
     normalizeQueryParams,
-    sanitizeURL,
-    shouldNavigate
+    sanitizeURL
 } from '../utils';
 
 @Component({
@@ -72,6 +62,7 @@ import {
         DotSeoMetaTagsService,
         DotSeoMetaTagsUtilService,
         DotWorkflowsActionsService,
+        DotUVENgEvenHandlerService,
         {
             provide: WINDOW,
             useValue: window
@@ -100,15 +91,10 @@ export class DotEmaShellComponent implements OnInit {
 
     readonly #router = inject(Router);
     readonly #location = inject(Location);
-    readonly #cd = inject(ChangeDetectorRef);
     readonly #siteService = inject(SiteService);
     readonly #activatedRoute = inject(ActivatedRoute);
-    readonly #messageService = inject(MessageService);
-    readonly #dotPageApiService = inject(DotPageApiService);
-    readonly #dotMessageService = inject(DotMessageService);
     readonly #dialogService = inject(DotEditorDialogService);
-    readonly #dotContentletService = inject(DotContentletService);
-    readonly #dotHttpErrorManagerService = inject(DotHttpErrorManagerService);
+    readonly #dotNgEventHandlerService = inject(DotUVENgEvenHandlerService);
 
     protected readonly $shellProps = this.uveStore.$shellProps;
 
@@ -246,262 +232,13 @@ export class DotEmaShellComponent implements OnInit {
         this.#location.go(urlTree.toString());
     }
 
-    protected handleNgEvent({ event, actionPayload, clientAction }: DialogAction) {
-        const { detail } = event;
-
-        switch (event.detail.name) {
-            case NG_CUSTOM_EVENTS.UPDATE_WORKFLOW_ACTION: {
-                const pageIdentifier = this.uveStore.pageAPIResponse().page.identifier;
-                this.uveStore.getWorkflowActions(pageIdentifier);
-                break;
-            }
-
-            case NG_CUSTOM_EVENTS.CONTENT_SEARCH_SELECT: {
-                const { pageContainers, didInsert } = insertContentletInContainer({
-                    ...actionPayload,
-                    newContentletId: detail.data.identifier
-                });
-
-                if (!didInsert) {
-                    this.handleDuplicatedContentlet();
-
-                    return;
-                }
-
-                this.uveStore.savePage(pageContainers);
-                break;
-            }
-
-            case NG_CUSTOM_EVENTS.CREATE_CONTENTLET: {
-                this.#dialogService.createContentlet({
-                    contentType: detail.data.contentType,
-                    url: detail.data.url,
-                    actionPayload
-                });
-
-                this.#cd.detectChanges();
-                break;
-            }
-
-            case NG_CUSTOM_EVENTS.CANCEL_SAVING_MENU_ORDER: {
-                this.#dialogService.resetDialog();
-                this.#cd.detectChanges();
-                break;
-            }
-
-            case NG_CUSTOM_EVENTS.SAVE_MENU_ORDER: {
-                this.#messageService.add({
-                    severity: 'success',
-                    summary: this.#dotMessageService.get(
-                        'editpage.content.contentlet.menu.reorder.title'
-                    ),
-                    detail: this.#dotMessageService.get('message.menu.reordered'),
-                    life: 2000
-                });
-
-                this.uveStore.reloadCurrentPage();
-                this.#dialogService.resetDialog();
-                break;
-            }
-
-            case NG_CUSTOM_EVENTS.LANGUAGE_IS_CHANGED: {
-                const htmlPageReferer = event.detail.payload?.htmlPageReferer;
-                const url = new URL(htmlPageReferer, window.location.origin); // Add base for relative URLs
-                const targetUrl = getTargetUrl(
-                    url.pathname,
-                    this.uveStore.pageAPIResponse().urlContentMap
-                );
-                const language_id = url.searchParams.get('com.dotmarketing.htmlpage.language');
-
-                if (shouldNavigate(targetUrl, this.uveStore.pageParams().url)) {
-                    // Navigate to the new URL if it's different from the current one
-                    this.uveStore.loadPageAsset({ url: targetUrl, language_id });
-
-                    return;
-                }
-
-                this.uveStore.loadPageAsset({
-                    language_id
-                });
-
-                break;
-            }
-
-            case NG_CUSTOM_EVENTS.ERROR_SAVING_MENU_ORDER: {
-                this.#messageService.add({
-                    severity: 'error',
-                    summary: this.#dotMessageService.get(
-                        'editpage.content.contentlet.menu.reorder.title'
-                    ),
-                    detail: this.#dotMessageService.get(
-                        'error.menu.reorder.user_has_not_permission'
-                    ),
-                    life: 2000
-                });
-
-                break;
-            }
-
-            case NG_CUSTOM_EVENTS.FORM_SELECTED: {
-                const formId = detail.data.identifier;
-
-                this.#dotPageApiService
-                    .getFormIndetifier(actionPayload.container.identifier, formId)
-                    .pipe(
-                        tap(() => {
-                            this.uveStore.setUveStatus(UVE_STATUS.LOADING);
-                        }),
-                        map((newFormId: string) => {
-                            return {
-                                ...actionPayload,
-                                newContentletId: newFormId
-                            };
-                        }),
-                        catchError(() => EMPTY),
-                        take(1)
-                    )
-                    .subscribe((response) => {
-                        const { pageContainers, didInsert } = insertContentletInContainer(response);
-
-                        if (!didInsert) {
-                            this.handleDuplicatedContentlet();
-                            this.uveStore.setUveStatus(UVE_STATUS.LOADED);
-                        } else {
-                            this.uveStore.savePage(pageContainers);
-                        }
-                    });
-
-                break;
-            }
-
-            case NG_CUSTOM_EVENTS.SAVE_PAGE: {
-                const { shouldReloadPage, contentletIdentifier } = detail.payload ?? {};
-                const pageIdentifier = this.uveStore.pageAPIResponse().page.identifier;
-                const isGraphqlPage = !!this.uveStore.graphql();
-
-                if (shouldReloadPage) {
-                    this.reloadURLContentMapPage(contentletIdentifier);
-
-                    return;
-                }
-
-                if (clientAction === CLIENT_ACTIONS.EDIT_CONTENTLET || !isGraphqlPage) {
-                    this.notifyContentOutsidePageHasChanged();
-                }
-
-                if (contentletIdentifier === pageIdentifier || !actionPayload) {
-                    this.handleReloadPage(event);
-
-                    return;
-                }
-
-                this.handleContentSave(event, actionPayload);
-
-                break;
-            }
-        }
-    }
-
     /**
-     * Handles the save page event triggered from the dialog.
+     * Handle the event triggered from the dialog.
      *
-     * @param {CustomEvent} event - The event object containing details about the save action.
-     * @return {void}
+     * @param {DialogAction} event - The event object containing details about the action.
+     * @memberof DotEmaShellComponent
      */
-    private handleReloadPage(event: CustomEvent): void {
-        // Move this to the GetTargetUrl function
-        const htmlPageReferer = event.detail.payload?.htmlPageReferer;
-        const url = new URL(htmlPageReferer, window.location.origin);
-        const targetUrl = getTargetUrl(url.pathname, this.uveStore.pageAPIResponse().urlContentMap);
-        // END
-
-        if (shouldNavigate(targetUrl, this.uveStore.pageParams().url)) {
-            // Navigate to the new URL if it's different from the current one
-            this.uveStore.loadPageAsset({ url: targetUrl });
-
-            return;
-        }
-
-        this.uveStore.reloadCurrentPage();
-    }
-
-    private handleContentSave(event: CustomEvent, actionPayload) {
-        const newContentletId = event.detail.payload?.newContentletId ?? '';
-
-        const { pageContainers, didInsert } = insertContentletInContainer({
-            ...actionPayload,
-            newContentletId
-        });
-
-        if (!didInsert) {
-            this.handleDuplicatedContentlet();
-
-            return;
-        }
-
-        this.uveStore.savePage(pageContainers);
-    }
-
-    private handleDuplicatedContentlet() {
-        this.#messageService.add({
-            severity: 'info',
-            summary: this.#dotMessageService.get('editpage.content.add.already.title'),
-            detail: this.#dotMessageService.get('editpage.content.add.already.message'),
-            life: 2000
-        });
-
-        this.uveStore.resetEditorProperties();
-        this.#dialogService.resetDialog();
-    }
-
-    /**
-     * Reload the URL content map page
-     *
-     * @private
-     * @param {string} inodeOrIdentifier
-     * @memberof EditEmaEditorComponent
-     */
-    private reloadURLContentMapPage(inodeOrIdentifier: string): void {
-        // Set loading state to prevent the user to interact with the iframe
-        this.uveStore.setUveStatus(UVE_STATUS.LOADING);
-
-        this.#dotContentletService
-            .getContentletByInode(inodeOrIdentifier)
-            .pipe(
-                catchError((error) => this.handlerError(error)),
-                filter((contentlet) => !!contentlet)
-            )
-            .subscribe(({ URL_MAP_FOR_CONTENT }) => {
-                if (URL_MAP_FOR_CONTENT != this.uveStore.pageParams().url) {
-                    // If the URL is different, we need to navigate to the new URL
-                    this.uveStore.loadPageAsset({ url: URL_MAP_FOR_CONTENT });
-
-                    return;
-                }
-
-                // If the URL is the same, we need to fetch the new page data
-                this.uveStore.reloadCurrentPage();
-            });
-    }
-
-    /**
-     * Handle the error
-     *
-     * @private
-     * @param {HttpErrorResponse} error
-     * @return {*}
-     * @memberof EditEmaEditorComponent
-     */
-    private handlerError(error: HttpErrorResponse) {
-        this.uveStore.setEditorState(EDITOR_STATE.ERROR);
-
-        return this.#dotHttpErrorManagerService.handle(error).pipe(map(() => null));
-    }
-
-    private notifyContentOutsidePageHasChanged() {
-        // this.contentWindow?.postMessage(
-        //     { name: __DOTCMS_UVE_EVENT__.UVE_RELOAD_PAGE },
-        //     "*"
-        // );
+    protected handleNgEvent(event: DialogAction) {
+        this.#dotNgEventHandlerService.handleNgEvent(event);
     }
 }
